requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.0
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
httpie==3.2.2
idna==3.4
importlib-metadata==6.7.0
ipython==8.14.0
jedi==0.18.2
Markdown==3.4.3
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
openai==0.27.8
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
PySocks==1.7.1
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
